# dev/openapi.yaml
openapi: 3.0.3
info:
  title: Packing List API
  version: 0.1.0
  description: Backpacking gear management API
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/labels:
    get:
      tags:
        - labels
      summary: List all labels
      operationId: listLabels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - labels
      summary: Create a new label
      operationId: createLabel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelCreate'
      responses:
        '201':
          description: Label created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/packing-lists:
    get:
      tags:
        - packing-lists
      summary: List all packing lists
      operationId: listPackingLists
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of packing lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackingList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - packing-lists
      summary: Create a new packing list
      operationId: createPackingList
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackingListCreate'
      responses:
        '201':
          description: Packing list created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackingList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/packing-lists/{listId}:
    get:
      tags:
        - packing-lists
      summary: Get packing list by ID
      operationId: getPackingListById
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Packing list details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackingList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - packing-lists
      summary: Update packing list
      operationId: updatePackingList
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackingListUpdate'
      responses:
        '200':
          description: Packing list updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackingList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - packing-lists
      summary: Delete packing list
      operationId: deletePackingList
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Packing list deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/packing-lists/{listId}/labels:
    get:
      tags:
        - packing-list-labels
      summary: List labels for a packing list
      operationId: listPackingListLabels
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of packing list labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - packing-list-labels
      summary: Add label to packing list
      operationId: addPackingListLabel
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackingListLabelAdd'
      responses:
        '201':
          description: Label added to packing list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/packing-lists/{listId}/labels/{labelId}:
    delete:
      tags:
        - packing-list-labels
      summary: Remove label from packing list
      operationId: removePackingListLabel
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: labelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Label removed from packing list
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
...
    Label:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          nullable: true
          description: Hex color code for UI display (e.g., #FF5733)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LabelCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
          description: Hex color code for UI display (e.g., #FF5733)
...
    PackingList:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PackingListCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    PackingListUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    PackingListLabelAdd:
      type: object
      required:
        - label_id
      properties:
        label_id:
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT